requirements.txt
Django~=4.1.6
setuptools~=63.4.1
cryptography~=37.0.1
keyring~=23.4.0
django-phone-field~=1.8.1
ruamel-yaml
